refactor: Websocket내에서 사용하는 jwtservice 위치 수정
